Map security from old organization to new organization (#46)
* feat: initial version of mapping code
I contain the initial version of the user mapping
script, which will map old GHEC users to new GHEC
users, using a CSV file. I am called as part of the
larger copy-permissions-between-orgs-repos.sh script.

* fix: set the mapping file name in the script.

Instead of passing in the mapping file name
set it in the script. This allows these changes
to more easily work with the other scripts, in
case someone creates a different mapping script.

* fix: made file executable

* fix: pathed to csv file

* feat: created script to move users after migration

* initial version of post migration script

* chore: updated docs

* Final version of security scripts

* removed unneeded readme

* Update gh-cli/get-organization-members.sh

* chore: rename path

* docs: add small readme

* fix: minor tweaks

---------

Co-authored-by: Josh Johanning <joshjohanning@github.com>